+++ cluster/admin.lua
@@ -151,12 +151,6 @@ local function get_servers_and_replicasets()
             end
         end
 
-        for role, enabled in pairs(enabled_roles) do
-            if enabled and not utils.table_find(known_roles, role) then
-                table.insert(replicasets[replicaset_uuid].roles, role)
-            end
-        end
-
         if replicaset.roles['vshard-storage'] then
             replicasets[replicaset_uuid].weight = replicaset.weight or 0.0
         end